+++ src/App.tsx
@@ -20,7 +20,7 @@ import { useThemeContext } from './theme/ThemeContext';
 import { SharedChallengeView } from './components/creator/SharedChallengeView';
 import { PilasBloquesApi } from './pbApi';
 import { Ember } from './emberCommunication';
-import { ChallengeView } from './components/ChallengeView/ChallengeView';
+import { ChallengeView } from './components/challengeView/ChallengeView';
 
 const AnalyticsComponent = () => {
   const location = useLocation();
