--- src/components/ChallengeView/Scene.tsx
@@ -1,5 +1,0 @@
-import { PBCard } from "../PBCard"
-
-export const Scene = () => {
-    return <PBCard sx={{width: "400px"}}>LA ESCENA</PBCard>
-}
