+++ src/components/challengeView/Scene.tsx
@@ -1,0 +1,6 @@
+import { PBCard } from "../PBCard"
+export const Scene = () => {
+    return <PBCard sx={{width: "400px"}}>
+        LA ESCENA
+    </PBCard>
+}
