+++ src/components/challengeView/SceneButtons.tsx
