+++ src/components/challengeView/StatementDescription.tsx
